Working on Buffer Math implementation
Seems to be working, but now EC multiplication (specifically the ECPointFp.twice method) is now very slow. Using more "modQ" calls sped it up, but it still grinds to a standstill after a few calls to "twice"; need to find if there's a memory leak or some other inefficiency there.